Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int a;
-    scanf("%d",&a);
+    float a;
+    scanf("%f",&a);
     float PI=3.14;
-    printf("Area: %.2d",PI*a*a);
+    printf("Area: %.2f",PI*a*a);
     return 0;
 }
